// src/types/Product.ts
export interface ProductType {
        productName: string
        productPrice: number
        productCount: number
        discountPrice : number,
        productCategory: number
        productSize : string
        productCurrency: string 
        productLabel : string,
        productDescription: string,
        primaryImageUrl: string,
        cartImgUrl: string,
        hoverImgUrl: string
}